+++ authentication/cdk/models.py
@@ -4,8 +4,6 @@ from typing import TypedDict
 class Context(TypedDict):
     environment_name: str
     mtls_subdomain: str
-    mtls_certificate: str
     subdomain: str
-    certificate: str
     hosted_zone_name: str
     scheme_base_url: str